Changes to support improved logging of canoe-parallel.sh -lb-by-sent.
Affects how logs are output in -worker-cmd and -unordered-cat modes.
In particular, makes sure time-mem monitoring output and RP-Totals are saved
even when -worker-cmd is used.

Adds -quiet-daemon option to run-parallel.sh, for silencing the daemon but not
the rest of run-parallel.sh

Also, daemon, in the computer sense and the supernatural being sense, is spelled
with the a before the e, or sometimes without the a, but never with the a after
the e, despite the 8 millions hits Google reports for the wrong spelling:
 - http://en.wikipedia.org/wiki/Daemon_(computing)
 - http://catb.org/~esr/jargon/html/D/daemon.html
 - http://www.oed.com/view/Entry/249688 (computer daemon or demon)
 - http://www.oed.com/view/Entry/49788 (demon or daemon: supernat. being or evil spirit)