Prepend all fatal error messages from Perl scripts with "Error: ".
This is to make it easier for the user to find errors in log files, e.g.
using grep.

Likewise, warning messages start with "Warning: ".

Always print an error message if exiting early with an error code.

Also, fixed some top comment blocks and/or help messages along the way.

(This commit is based on changes in the PORTAGEshared commit:
9f391f09f8ea6ca843c1c6f1d875a2289b1f759e)