Don't use PID for workdir name uniqueness; PID not unique on GPSC.
On the GPSC, PID is not unique across containers, so commands run in
different containers may have the same PID. This happens when jobs are
scheduled using psub.

As a consequence our scripts should not use the PID as a unique
identifier on workdirs or temp files. Use tempdir or mktemp instead.