run-parallel.sh on GPSC: CPU requirements inheritance; mem constraints
As it did on Balzac, run-parallel.sh now propagates the number of CPUS to children unless
the -psub option explicitly specifies different requirements.

As it did on Balzac, we make sure there's enough memory for each local worker.

And as on Balzac, memory requests for the master are not propagated, they're only used as
contraints.