run-parallel.sh: fix detection of -c make calls.
Code to detect if command is make should only trigger when command starts with make. I
guess in previous installations, << if [[ $* =~ ^make ]] >> must have checked for $* as a
whole string, but now it seems I have to use "$*" to get the same result.

This patch is to fix the filter_models test suite in Portage R&D.