patch run-parallel to minize the number of calls to jobst
With the last change to jobst, it no longer exits with a non-zero code when the workers
are all done, so it would systemcally call jobst 40 times at the end of each job - not
cluster friendly!

Now, we just a grep expression to test if the workers are still running instead.

Also, wait 3 seconds between each test for running workers, since things don't exit that
fast.