Made psub and run-parallel.sh more robust w.r.t. job names and jobsub failures.
In particular, double quote the jobname variable when using it in
commands to better handle special characters. This is paricularly
important to avoid getting:
   sh: -c: line 0: unexpected EOF while looking for matching `''
   sh: -c: line 1: syntax error: unexpected end of file
when there is a single quote in the jobname. [Note: framelab was
sending jobnames surrounded by single quotes. When psub truncated
a long jobname, it was left with an unmatched single quote in the
name resulting in the above cryptic error message.]

Even more importantly, I changed psub to check the return code
of the jobsub invocation and to error_exit upon failure after
outputting an informative message. This should fix some of our
previously mysterious problems with run-parallel.sh and parallelize.pl.